chore: add nix dev environment (#86)
* WIP attempt at running nix flake

* add git to fix issue with openssl 3 linking

* node setup

* Nix flakes for stump

* fix github actions

* fix github actions

* add comment

* nix flake in ci

* comments

* hopefully fix nix ci build

* oups

* comments

* Update build_nix.yml

* Update build_nix.yml

* Update build_nix.yml

* Update build_nix.yml

+++ apps/server/src/routers/api/library.rs
@@ -178,7 +178,7 @@ async fn get_library_series(
 			let media_count = match media_counts.get(&s.id) {
 				Some(count) => count.to_owned(),
 				_ => 0,
-			} as i64;
+			};
 
 			(s.to_owned(), media_count).into()
 		})
